+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All notable changes to this project will be documented in this file.
   z-positions of localizations with x and y PSF widths.
 - A 3D astigmatic imaging example notebook was added to the *examples*
   folder to explain how to use the new functionality.
+- Wobble curves in 3D astigmatic imaging are also taken accounted for
+  using the `CalibrateAstigmatism` and `ComputeZPosition` processors.
 
 ### Changed
 - The version number contained in the `bstore.__version__` string is